address sdk module review
